// frontend/src/app/models/teacher.ts
export interface Teacher {

    firstName: string ;
    lastName: string;
    gender: string;
    mobile: string;
    designation: string;
    departments: string;
    email: string;
    address: string;
    education:string;
    id: any;
    degree:string;
    joiningDate:any;
    profilePhoto:string;
  }